+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added serialization and equality comparison for `TransactionScript` (#824).
 - [BREAKING] Migrated to Miden VM v0.10 (#826).
 - Added conversions for `NoteExecutionHint` (#827).
+- [BREAKING] Refactored account storage and vault deltas (#822).
 
 ## 0.4.0 (2024-07-03)